pyproject.toml
[project]
name = "nonebot-plugin-pcr-sign"
version = "0.3.0"
description = "nonebot_plugin_hoshino_sign 的重置版"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.2",
    "nonebot-plugin-alconna>=0.53.1",
    "nonebot-plugin-argot>=0.1.5",
    "nonebot-plugin-htmlrender>=0.4.0",
    "nonebot-plugin-localstore>=0.7.2",
    "nonebot-plugin-orm>=0.7.6",
    "nonebot-plugin-uninfo>=0.4.1",
    "nonebot2>=2.4.0",
]

[project.entry-points.nb_scripts]
pcr = "nonebot_plugin_pcr_sign.cli.__main__:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.optional-dependencies]
adapters = [
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-adapter-telegram>=0.1.0b20",
]
test = [
    "nonebot2[fastapi,httpx]>=2.4.0",
    "nonebot-plugin-orm[default]>=0.7.6",
]
cli = [
    "aiosqlite>=0.20.0",
    "cli-lite>=0.10.2",
]

[tool.nonebot]
plugins = ["nonebot_plugin_pcr_sign"]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
    { name = "Telegram", module_name = "nonebot.adapters.telegram" }
]

[tool.black]
line-length = 90
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 90
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 90
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901"]

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "basic"
